+++ CHANGELOG.md
@@ -60,7 +60,8 @@
 * nostr: add `TagStandard::GitMaintainers` and `TagKind::Maintainers` variants ([Yuki Kishimoto])
 * nostr: add `TagStandard::Web` and `TagKind::Web` variants ([Yuki Kishimoto])
 * nostr: add `EventBuilder::git_repository_announcement` ([Yuki Kishimoto])
-* nostr: add `EventBuilder::git_issue` ([Yuki Kishimoto]
+* nostr: add `EventBuilder::git_issue` ([Yuki Kishimoto])
+* nostr: add `Tag::reference` constructor ([Yuki Kishimoto])
 * database: add `Backend::is_persistent` method ([Yuki Kishimoto])
 * database: add `Events` struct ([Yuki Kishimoto])
 * relay-builder: add `LocalRelay` and `RelayBuilder` ([Yuki Kishimoto])